# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SchoolPilot API
  description: |
    Multi-tenant school management SaaS API.
    Authenticated routes require `Authorization: Bearer <accessToken>`.
    OpenAPI spec also served at `GET /api/docs/openapi.yaml`.
  version: 1.0.0
  contact:
    name: Great Goodluck Schools
servers:
  - url: http://localhost:4000
    description: Local development
  - url: https://api.yourschoolpilot.com
    description: Production
...
paths:
...
  /api/public/home:
    get:
      tags: [Public]
      summary: Public website home data
      parameters:
        - name: schoolId
          in: query
          schema: { type: string }
      responses:
        '200':
          description: CMS content for landing page
...
  /api/system/status:
    get:
      tags: [Platform]
      summary: Integration readiness (SMTP, AI, payments)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Integration status list
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT